# 03-Azure/01-04-AI/01_Agentic_SDLC/src/api-ts/api-swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Express API with Swagger",
        "version": "1.0.0",
        "description": "REST API documentation using Swagger/OpenAPI"
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "Development server"
        }
    ],
    "paths": {
...
        "/api/orders": {
            "get": {
                "summary": "Returns all orders",
                "tags": [
                    "Orders"
                ],
                "responses": {
                    "200": {
                        "description": "List of all orders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Order"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "summary": "Create a new order",
                "tags": [
                    "Orders"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Order"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Order created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "get": {
                "summary": "Get an order by ID",
                "tags": [
                    "Orders"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Order ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            },
            "put": {
                "summary": "Update an order",
                "tags": [
                    "Orders"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Order ID"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Order"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Order updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Order"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete an order",
                "tags": [
                    "Orders"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Order ID"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Order deleted successfully"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            }
        },
        "/api/orders/branch/{branchId}/history": {
            "get": {
                "summary": "Get order history for a branch",
                "tags": [
                    "Orders"
                ],
                "parameters": [
                    {
                        "in": "path",
                        "name": "branchId",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "description": "Branch ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Branch orders with line items and totals",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrderHistoryItem"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid branch ID"
                    },
                    "404": {
                        "description": "Branch not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Order": {
                "type": "object",
                "required": [
                    "orderId",
                    "branchId",
                    "orderDate"
                ],
                "properties": {
                    "orderId": {
                        "type": "integer",
                        "description": "The unique identifier for the order"
                    },
                    "branchId": {
                        "type": "integer",
                        "description": "The ID of the branch that placed the order"
                    },
                    "orderDate": {
                        "type": "string",
                        "format": "date-time",
                        "description": "The date and time when the order was placed"
                    },
                    "name": {
                        "type": "string",
                        "description": "The name of the order"
                    },
                    "description": {
                        "type": "string",
                        "description": "Additional details about the order"
                    },
                    "status": {
                        "type": "string",
                        "description": "The current status of the order",
                        "enum": [
                            "cart",
                            "pending",
                            "processing",
                            "shipped",
                            "delivered",
                            "cancelled"
                        ]
                    }
                }
            },
...
            "OrderHistoryDetail": {
                "type": "object",
                "required": [
                    "orderDetailId",
                    "productId",
                    "productName",
                    "quantity",
                    "unitPrice",
                    "lineTotal"
                ],
                "properties": {
                    "orderDetailId": {
                        "type": "integer"
                    },
                    "productId": {
                        "type": "integer"
                    },
                    "productName": {
                        "type": "string"
                    },
                    "quantity": {
                        "type": "integer"
                    },
                    "unitPrice": {
                        "type": "number",
                        "format": "float"
                    },
                    "lineTotal": {
                        "type": "number",
                        "format": "float"
                    },
                    "notes": {
                        "type": "string"
                    }
                }
            },
            "OrderHistoryItem": {
                "type": "object",
                "required": [
                    "orderId",
                    "branchId",
                    "orderDate",
                    "name",
                    "status",
                    "total",
                    "details"
                ],
                "properties": {
                    "orderId": {
                        "type": "integer"
                    },
                    "branchId": {
                        "type": "integer"
                    },
                    "orderDate": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "total": {
                        "type": "number",
                        "format": "float"
                    },
                    "details": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/OrderHistoryDetail"
                        }
                    }
                }
            },
...
        }
    },
...
}